Reorganized folder, split classes into separate files
